123/web/requirements.txt
Flask>=2.0.1
redis>=4.0.2
python-dotenv>=0.19.0